Enable Hiccup doctype helpers in templates
First we need to wrap parenthesis around the string that slurp reads
from the template file, so that we get a list of objects with
read-string. Then we need to process all objects in that list with
eval and html.

Finally we need to use the hiccup.page-helpers namespace, so that the
helper methods are available when the template code is evaluated.