proper cssgen handling
'read-cssgen' was just calling 'css' on the 'eval'ed contents of each file, but
this only lets us create single-rule files.  I modified it to handle *.cssgen
files as containing a Clojure expression that evaluates to an array of cssgen
rules.

As well, the assignment of 'template-string' in the 'template' function
was yielding 'content' in the case that '(= template :none)'.  I modified it
to yield 'c', as I believ this is the appropriate value.